v0.77.0: CI hygiene — clang-tidy cleanup + adaptive version_test
- Fixed 73 clang-tidy errors across 16 files (init-variables,
  cert-err33-c, reinterpret_cast, roundings, narrowings, header
  modernization). Every touched TU recompiles clean with
  -Wall -Wextra -Wpedantic -Werror.
- Rewrote tests/unit/version_test.cpp to verify runtime accessor
  matches compile-time constants instead of pinning literal 74.
  Fixes coverage job which had been red since v0.75.0.
- Added docs/first_enable_verification.md — manual checklist for
  the v0.75.2 first-enable restart dialog.

No engine behavior changes, no binding changes, no new features.
v0.78.0 will pick up eviction tie-breaking; v0.79.0 distance-weighted
eviction.

+++ src/audio_engine/assets/audio_asset_registry.cpp
@@ -72,7 +72,8 @@ bool DecodeFully(IAudioDecoder& dec, std::vector<float>& out) {
         if (got == 0) break;
         out.insert(out.end(),
                     buf.begin(),
-                    buf.begin() + static_cast<size_t>(got) * channels);
+                    buf.begin() + static_cast<std::ptrdiff_t>(
+                        static_cast<size_t>(got) * channels));
         if (got < kChunk) break;
     }
     // v0.66.0: zero-frame decode is a failure. Previously returned
